yolo-service/requirements.txt
torch
ultralytics
boto3
pika
requests
python-dotenv